pyproject.toml
[tool.poetry]
name = "electronics-inventory-backend"
version = "0.1.0"
description = "Flask backend for hobby electronics parts inventory management system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.scripts]
cli = "app.cli:main"
dev = "run:main"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.0"
alembic = "^1.13.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
spectree = "^1.2.0"
flask-cors = "^4.0.0"
waitress = "^3.0.0"
boto3 = "^1.34.0"
celery = "^5.3.0"
openai = "^1.10.0"
anthropic = "^0.42.0"
paste = "^3.10.1"
dependency-injector = "^4.48.1"
pillow = "^10.2.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
python-magic = "^0.4.27"
validators = "^0.35.0"
prometheus-flask-exporter = "^0.23.0"
reportlab = "^4.0.0"
httpx = "^0.28.1"
pyjwt = "^2.11.0"
cryptography = "^46.0.4"
itsdangerous = "^2.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
mypy = "^1.8.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
types-flask = "^1.1.0"
types-flask-cors = "^4.0.0"
types-requests = "^2.31.0"
types-beautifulsoup4 = "^4.12.0"
types-pillow = "^10.2.0"
boto3-stubs = {extras = ["s3"], version = "^1.40.25"}
types-waitress = "^3.0.1.20250801"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
plugins = ["sqlalchemy.ext.mypy.plugin"]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_errors = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.models.*"
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v -m 'not integration'"
markers = [
    "integration: marks tests as integration tests (requires external dependencies)",
    "slow: marks tests as slow running tests"
]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::pluggy._warnings.PluggyTeardownRaisedWarning",
    "ignore::DeprecationWarning:pydantic.*"
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]
